% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_coverage.R
\name{bins_coverage}
\alias{bins_coverage}
\title{Bins coverage}
\usage{
bins_coverage(bins_gr, chr, strand, bam_content)
}
\arguments{
\item{bins_gr}{Bins positions in genomic coordinates as a GRanges}

\item{chr, strand}{Genomic position of the transcript (to filter the BAM)}

\item{bam_content}{Content of the BAM file containing reads}
}
\value{
Count of reads in bins
}
\description{
Bins coverage
}
