% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_coords_conversions.R
\name{bins_to_granges}
\alias{bins_to_granges}
\title{Convert bins to a list of GRanges objects in genomic coordinates}
\usage{
bins_to_granges(
  tx_chr,
  tx_strand,
  tx_start,
  tx_end,
  exons_lengths,
  exons_starts,
  bins,
  opt
)
}
\arguments{
\item{tx_chr, tx_strand, tx_start, tx_end}{Genomic position of the transcript}

\item{exons_lengths, exons_starts}{Structure of the transcript}

\item{bins}{Breaks for the bins}

\item{opt}{List containing endedness (do we count from 3' or from 5')}
}
\value{
A list of GRranges objects, each element of the list is one bin, There might be several Ranges within a single element if that bins spans several exons (making holes for introns)
}
\description{
Convert bins to a list of GRanges objects in genomic coordinates
}
\examples{
bins_to_granges(tx_chr = "I", tx_strand = "+",
  tx_start = 100, tx_end = 300,
  exons_lengths = c(10,50),exons_starts = c(0,150),
  bins = c(0,5,15,30),
  opt = list(endedness = 5L))
}
