% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_coverage.R
\name{breaks_coverage}
\alias{breaks_coverage}
\title{Breakpoints coverage}
\usage{
breaks_coverage(tx_chr, tx_strand, breaks_in_genomic, reads)
}
\arguments{
\item{tx_chr, tx_strand}{Chromosome and strand of the transcript}

\item{breaks_in_genomic}{Breakpoint positions in genomic coordinates}

\item{reads}{Content of the BAM file containing reads}
}
\value{
Count of reads on breakpoints
}
\description{
Breakpoints coverage
}
