% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_coords_conversions.R
\name{breaks_to_genomic}
\alias{breaks_to_genomic}
\title{Convert bins to a list of GRanges objects in genomic coordinates}
\usage{
breaks_to_genomic(
  tx_chr,
  tx_strand,
  tx_start,
  tx_end,
  spliced_tx_width,
  exons_lengths,
  exons_starts,
  breakpoints,
  opt
)
}
\arguments{
\item{tx_chr, tx_strand, tx_start, tx_end, spliced_tx_width}{Genomic position of the transcript}

\item{exons_lengths, exons_starts}{Structure of the transcript}

\item{breakpoints}{Coordinates of the breakpoints (in the spliced transcript)}

\item{opt}{List containing endedness (do we count from 3' or from 5')}
}
\value{
A list of GRranges objects, each element of the list is one breakpoint.
}
\description{
As input, we take the breakpoints, as output, we get GRanges that cover these points (1 bp-wide each).
 Note this is different from `bins_to_granges()` which returns the bins between the breakpoints.
}
\examples{
breaks_to_genomic(tx_chr = "I", tx_strand = "+",
  tx_start = 100, tx_end = 200, spliced_tx_width = 60,
  exons_lengths = c(10,50), exons_starts = c(0,150),
  breakpoints = c(0,5,15,30),
  opt = list(endedness = 5L))
}
