% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_coords_conversions.R
\name{coords_as_bin_granges}
\alias{coords_as_bin_granges}
\title{Coordinates to GRanges}
\usage{
coords_as_bin_granges(
  tx_chr,
  bins,
  tx_strand,
  tx_start,
  tx_ex_starts,
  tx_ex_len,
  opt
)
}
\arguments{
\item{tx_chr, tx_strand, tx_start}{Transcript genomic position}

\item{bins}{Bins positions in genomic coordinates}

\item{tx_ex_starts, tx_ex_len}{Starts and lengths of the exons}

\item{opt}{List containing endedness (do we count from 3' or from 5')}
}
\value{
A list of GRranges objects, each element of the list is one bin, There might be several Ranges within a single element if that bins spans several exons (making holes for introns)
}
\description{
Coordinates to GRanges
}
