% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_coords_conversions.R
\name{coords_as_min_granges}
\alias{coords_as_min_granges}
\title{Coordinates as minimal GRanges}
\usage{
coords_as_min_granges(tx_chr, tx_strand, breaks_in_genomic)
}
\arguments{
\item{tx_chr, tx_strand}{Transcript genomic position}

\item{breaks_in_genomic}{Breakpoints positions in genomic coordinates}
}
\value{
A Granges object, each range is 1 bp wide around the breakpoint
}
\description{
Coordinates as minimal GRanges
}
