% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_coords_conversions.R
\name{coords_to_genomic}
\alias{coords_to_genomic}
\title{Convert unspliced to genomic coordinates}
\usage{
coords_to_genomic(tx_st, tx_end, bins, tx_strand, opt)
}
\arguments{
\item{tx_st}{Genomic position of transcript start}

\item{tx_end}{Genomic position of transcript end}

\item{bins}{Bins positions in unspliced coordinates}

\item{tx_strand}{Genomic strand of transcript}

\item{opt}{List containing endedness (do we count from 3' or from 5')}
}
\value{
Bins positions in genomic coordinates
}
\description{
Convert unspliced to genomic coordinates
}
