% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_coords_conversions.R
\name{coords_to_unspliced}
\alias{coords_to_unspliced}
\title{Convert coordinates from spliced to unspliced}
\usage{
coords_to_unspliced(
  coords_in_spliced,
  tx_strand,
  exons_lengths,
  exons_starts,
  opt
)
}
\arguments{
\item{coords_in_spliced}{Bins positions in spliced coordinates (where introns count 0 bp)}

\item{tx_strand}{Genomic strand of this transcript}

\item{exons_lengths, exons_starts}{Exons lengths and starts}

\item{opt}{List containing endedness (do we count from 3' or from 5')}
}
\value{
Bins positions in unspliced coordinates (i.e. the transcript starts at 0, but we add space for the introns)
}
\description{
Convert coordinates from spliced to unspliced
}
