% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_coords_conversions.R
\name{last_element}
\alias{last_element}
\title{Select last element of vector}
\usage{
last_element(x)
}
\arguments{
\item{x}{Vector}
}
\value{
The last element of `x`
}
\description{
Select last element of vector
}
