% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_coverage.R
\name{breaks_coverage}
\alias{breaks_coverage}
\title{Breakpoints coverage}
\usage{
breaks_coverage(breaks_gr, reads)
}
\arguments{
\item{breaks_gr}{Breakpoint positions as a GRanges}

\item{reads}{Content of the BAM file containing reads}
}
\value{
Count of reads on breakpoints
}
\description{
Breakpoints coverage
}
